feat: AuthorizationDirective updates * Add `AbstractIfThenElseDirective` to reduce boilerplate code for structural directives * Break up `AuthorizationDirective` into multiple specific directives * Add support for `and/or`ing directive with addition condition
